Make "back" link work in more situations
The link will now work when there is no referrer (because the user typed
a URL in manually) but there is a history, and thus a page to go back
to. In this case, the URL shown in the href attribute will just be the
current page (since we can't access the URL from the history), but
clicking on the link will still go back a page in the history.