Adapt `user` table schema to work with OAuth flow
* Removed 'email' column.
* Added 'oauth_provider' and 'provider_user_id' columns.
* Added a unique constraint on ('oauth_provider', 'provider_user_id').

Additionally, the previous database initialization script created by
flask-migrate was updated. In development, we user sqlite3, which does
not support dropping columns from existing tables. This commit required
the removal of the 'email' column, which was previously declared NOT
NULL, so this would create issues in the future, even with brand-new
initialized databases, since *all* of the flask-migrate scripts would
still be run, including the one defining the original 'email' column.
Therefore, this commit also comments out the lines which created that
column in the earlier db initialization script.