Fix clickable area and display of intro links
The 'My Instruments' and 'Create New Instrument' links at the top of the
home page were clickable across the screen from their visual location.
This commit prevents that surprising behavior.

These links are now hidden at the largest breakpoint, since the same
links are provided in the nav bar, which is always visible after that
breakpoint.